third_party/WebKit/LayoutTests/imported/wpt/encoding/textdecoder-utf16-surrogates-expected.txt

This is a testharness.js-based test.
FAIL utf-16le - lone surrogate lead assert_equals: expected "\ufffd" but got "í €"
FAIL utf-16le - lone surrogate lead (fatal flag set) assert_throws: function "function () {
            new TextDecoder(t.encoding, {fa..." did not throw
FAIL utf-16le - lone surrogate trail assert_equals: expected "\ufffd" but got "í°€"
FAIL utf-16le - lone surrogate trail (fatal flag set) assert_throws: function "function () {
            new TextDecoder(t.encoding, {fa..." did not throw
FAIL utf-16le - unmatched surrogate lead assert_equals: expected "\ufffd\0" but got "í €\0"
FAIL utf-16le - unmatched surrogate lead (fatal flag set) assert_throws: function "function () {
            new TextDecoder(t.encoding, {fa..." did not throw
FAIL utf-16le - unmatched surrogate trail assert_equals: expected "\ufffd\0" but got "í°€\0"
FAIL utf-16le - unmatched surrogate trail (fatal flag set) assert_throws: function "function () {
            new TextDecoder(t.encoding, {fa..." did not throw
FAIL utf-16le - swapped surrogate pair assert_equals: expected "\ufffd\ufffd" but got "í°€í €"
FAIL utf-16le - swapped surrogate pair (fatal flag set) assert_throws: function "function () {
            new TextDecoder(t.encoding, {fa..." did not throw
Harness: the test ran to completion.
